refactor: numbers
